불필요 코드 제거
