--- src/main/kotlin/mashup/sideproject/orderpay/exception/BaseException.kt
@@ -1,9 +1,0 @@
-package mashup.sideproject.orderpay.exception
-
-import org.springframework.http.HttpStatus
-
-open class BaseException(
-        val code: Int,
-        override val message: String,
-        val httpStatus: HttpStatus
-) : RuntimeException()
