--- src/main/kotlin/mashup/sideproject/orderpay/model/dto/OrderRequestDto.kt
@@ -1,8 +1,0 @@
-package mashup.sideproject.orderpay.model.dto
-
-import mashup.sideproject.orderpay.model.dto.iamport.BaseDto
-
-data class OrderRequestDto(
-    val productIdList: List<Long> = ArrayList(),
-    val optionIdList: List<Long> = ArrayList()
-) : BaseDto
