+++ src/main/kotlin/mashup/sideproject/orderpay/model/dto/order/MerchantUidDto.kt
@@ -1,4 +1,4 @@
-package mashup.sideproject.orderpay.model.dto
+package mashup.sideproject.orderpay.model.dto.order
 
 import mashup.sideproject.orderpay.model.dto.iamport.BaseDto
